Finish generate_youtubeapi_testcase.rb
