+++ spec/fixtures/youtube_results.yml
@@ -1,5 +1,7 @@
 ---
-- id: ma67yOdMQfs
+- topics: surfing
+- index: 0
+  id: ma67yOdMQfs
   title: These Were The All-Time Surfing Moments Of The Year | Best Of 2020
   description: "Well, that was a weird ride. Though it hasn't been easy, at least
     when we fixed our gaze on the ocean — or favorite place in the world – very little
@@ -66,7 +68,22 @@
   - extreme athletes
   channelTitle: Red Bull Surfing
   channelId: UC--3c8RqSfAqYBdDjIG3UNA
-- id: A5oD6RhwCi0
+- index: 1
+  id: vD6WGEkw8Jc
+  title: Mason & Coco Ho Surfing Backdoor
+  description: "It’s that time of year again! The beautiful North Shore of Oahu bringing
+    everybody together to surf. \n\nIn this video Mason & Coco Ho surf with the Rip
+    Curl team plus some world travelers and local faces. \n\nLocation: Backdoor &
+    Pipeline with a lot more sand than usual.\n\nSurfers: Mason Ho, Coco Ho, Crosby
+    Colapinto, Matthew McGillivray Morgan Ciblic, Mike O’Shaughnessy, Nalu Deodato,
+    Luke Tema, Yago Dora, Leo Fioravanti , Thatcher Johnson, Zoe McDougal, Luke Swanson
+    & Joey Johnston. \n\nJaMs: mom collection Carlos S.\n\nOnly some surfboard info
+    in end credit roll plus DING reporting for duty."
+  tags:
+  channelTitle: Ho & Pringle Productions
+  channelId: UCeYue9Nbodzg3T1Nt88E3fg
+- index: 2
+  id: A5oD6RhwCi0
   title: Points South with Dave Rastovich
   description: |-
     First release of a surf trip I took with Dave Rastovich and Dan Ross to explore some waves in the Southern Ocean.
@@ -118,7 +135,8 @@
   - barrels
   channelTitle: John John Florence
   channelId: UCNyxrRXa3PvyJYuFP51iU-Q
-- id: wZYeAZtJUnE
+- index: 3
+  id: wZYeAZtJUnE
   title: Huge Wave Slams Unlucky Surfers - Uluwatu
   description: "This was one of the biggest days of the year a few months ago. Uluwatu's
     Outside Corner had some bombs, but the lineup looked really unpredictable. \n\nThere
@@ -162,117 +180,42 @@
   - surfing
   channelTitle: Surfers of Bali
   channelId: UCzcQOTuXYGuCvTekySb_CeQ
-- id: VwQwnvILthU
-  title: Kolohe Andino Surfing Steamer Lane in Santa Cruz, California - October 26th
-    and 27th, 2021
-  description: Kolohe surfing big steamer lane one day then a smaller day after.
+- index: 4
+  id: aS-NoUsMcac
+  title: Ian Walsh's How-To Guide for a Big Wave Winter | FOLLOWING Ep7
+  description: "From Ian Walsh preparing his equipment for XXL waves, to giving his
+    boards their first test-drives in somewhat surprising surf, this installment of
+    Following is an insight into the thoroughly prepared existence of a man who leaves
+    nothing to chance. \n\nAs the Northern winter rolls around again we revisit last
+    year’s fly-on-the-wall edit from the crew at Inherent Bummer, who first met up
+    with Ian Walsh at home in Maui, before following him across the Pacific to surf
+    Mavericks in California, with good friends and fellow chargers Kai Lenny and Lucas
+    “Chumbo” Chianca.  \n\nIt’s an eye-opening look at what goes on in the world of
+    the modern big surfer, who live their lives at the whim of the weather and need
+    to be ready to drop everything and go at the first sign of a storm front. We’ve
+    all seen the dramatic images of surfing’s gladiators taming monsters in the middle
+    of the ocean, but have you seen the 4am wake-ups, the endless wrestling of huge
+    board bags through airports, and the shivering pre-dawn carpark rev up sessions?
+    Hit play above and enter the world of the wave warrior now.  \n\nRead more about
+    the wild world of big wave warrior Ian Walsh \U0001F447\U0001F3FB\nhttps://www.redbull.com/int-en/following-ian-walsh\n\nCheck
+    out every episode of Following:\nEp6: Caroline Marks - https://win.gs/3o9w0Tt\nEp5:
+    Eli Hanneman - https://win.gs/3H9HAqr\nEp4: Kanoa Igarashi - https://win.gs/3oi9cRp\nEp3:
+    Julian Wilson - https://win.gs/3kqoVNc\nEp2: Carissa Moore - https://win.gs/3kszgIC\nEp1:
+    Kolohe Andino - https://win.gs/3EXinh0\n\n_____________________________________________________________\r\n\r\nExperience
+    the world of Red Bull like you have never seen it before. With the best action
+    sports clips on the web and original series, prepare for your \"stoke factor\"
+    to be at an all time high.\r\n\r\nSubscribe to Red Bull Surfing on Youtube: https://win.gs/SubToRedBullSurfing\r\n\r\nGet
+    the FREE Red Bull TV apps for all your devices: http://onelink.to/yfbct7\r\n\r\nWatch
+    Red Bull TV: http://win.gs/WatchRBTV\r\n\r\nRed Bull on Facebook: http://win.gs/redbullfb\r\n\r\nRed
+    Bull on Instagram: http://win.gs/RedBullInstagram"
   tags:
-  - surfing
-  - world surf league
-  - wsl
+  - red bull surfing
+  - red bull
   - surf
-  - Santa Cruz
-  - big waves
-  - ocean
-  - drone
-  - dji
-  - Kolohe andino
-  - kelly slater
-  - honolua bay
-  - John John florence
-  - steamer lane
-  - hurley
-  - redbull
-  - oakley
-  - olympics
-  - team usa
-  - gold medal
-  - surfer
-  - sony
-  - edits
-  - viral
-  - tiktok
-  - instagram
-  - kim kardashian
-  channelTitle: Tucker Wooding Media
-  channelId: UCgl3a2mzUzgKqb6ey5MpP3w
-- id: Smgql_DESn8
-  title: Stray Kids "Surfin' (리노, 창빈, 필릭스)" Video
-  description: |-
-    Stray Kids(스트레이 키즈) "Surfin' (리노, 창빈, 필릭스)" Video
-
-    Stray Kids THE 2ND ALBUM "NOEASY"
-    iTunes & Apple Music: https://apple.co/3moxkm9
-    Spotify: https://spoti.fi/3Djh3VR
-
-    Stray Kids Official YouTube: https://www.youtube.com/c/StrayKids
-    Stray Kids Official Facebook: https://www.facebook.com/JYPEStrayKids/
-    Stray Kids Official Instagram: https://www.instagram.com/realstraykids/
-    Stray Kids Official Twitter: https://twitter.com/Stray_Kids
-    Stray Kids Official Fan's: https://fans.jype.com/StrayKids
-
-    #StrayKids #스트레이키즈
-    #Surfin #NOEASY
-    #소리꾼 #Thunderous
-    #YouMakeStrayKidsStay
-
-    Copyrights 2021 ⓒ JYP Entertainment. All Rights Reserved.
-  tags:
-  - Stray Kids
-  - 스트레이 키즈
-  - SKZ
-  - 스키즈
-  - 방찬
-  - BANG CHAN
-  - 리노
-  - 이민호
-  - LEE KNOW
-  - MINHO
-  - LEEMINHO
-  - 창빈
-  - 서창빈
-  - CHANGBIN
-  - SEOCHANGBIN
-  - 현진
-  - 황현진
-  - HYUNJIN
-  - HWANGHYUNJIN
-  - 한
-  - 지성
-  - 한지성
-  - HAN
-  - HANJISUNG
-  - 필릭스
-  - FELIX
-  - 승민
-  - 김승민
-  - SEUNGMIN
-  - KIMSEUNGMIN
-  - 아이엔
-  - 양정인
-  - 정인
-  - I.N
-  - YANGJEONGIN
-  - JEONGIN
-  - 소리꾼
-  - Thunderous
-  - NOEASY
-  - 스트레이 키즈 NOEASY
-  - Stray Kids NOEASY
-  - 스트레이 키즈 소리꾼
-  - 스트레이 키즈 Thunderous
-  - 스키즈 소리꾼
-  - 스트레이 키즈 Surfin
-  - Surfin
-  - 스키즈 Surfin
-  - 스트레이 키즈 서핀
-  - Stray Kids Surfin
-  - Stray Kids 서핀
-  - 서핀 뮤직비디오
-  - 서핀 뮤비
-  - Surfin MV
-  - 서핀 MV
-  - Surfin music video
-  - 리노 창빈 필릭스
-  channelTitle: Stray Kids
-  channelId: UC9rMiEjNaCSsebs31MRDCRA
+  - surfing
+  - surfboarding
+  - big wave
+  - waves
+  - swell
+  channelTitle: Red Bull Surfing
+  channelId: UC--3c8RqSfAqYBdDjIG3UNA
