--- spec/fixtures/youtube_results_all_01.yml
@@ -1,148 +1,0 @@
----
-kind: youtube#searchListResponse
-etag: tt_p1FV-_8WYqriq_GazLxkW4Fk
-nextPageToken: CAUQAA
-regionCode: TW
-pageInfo:
-  totalResults: 1000000
-  resultsPerPage: 5
-items:
-- kind: youtube#searchResult
-  etag: U21Po-KMvdNZ7xiYZgqhPjinEps
-  id:
-    kind: youtube#video
-    videoId: ma67yOdMQfs
-  snippet:
-    publishedAt: '2021-01-23T17:00:15Z'
-    channelId: UC--3c8RqSfAqYBdDjIG3UNA
-    title: These Were The All-Time Surfing Moments Of The Year | Best Of 2020
-    description: Well, that was a weird ride. Though it hasn't been easy, at least
-      when we fixed our gaze on the ocean — or favorite place in the world – very
-      little had changed.
-    thumbnails:
-      default:
-        url: https://i.ytimg.com/vi/ma67yOdMQfs/default.jpg
-        width: 120
-        height: 90
-      medium:
-        url: https://i.ytimg.com/vi/ma67yOdMQfs/mqdefault.jpg
-        width: 320
-        height: 180
-      high:
-        url: https://i.ytimg.com/vi/ma67yOdMQfs/hqdefault.jpg
-        width: 480
-        height: 360
-    channelTitle: Red Bull Surfing
-    liveBroadcastContent: none
-    publishTime: '2021-01-23T17:00:15Z'
-- kind: youtube#searchResult
-  etag: RNQ7jRXz8VT-oxHUePgRCVEbanE
-  id:
-    kind: youtube#video
-    videoId: wZYeAZtJUnE
-  snippet:
-    publishedAt: '2021-11-11T11:00:24Z'
-    channelId: UCzcQOTuXYGuCvTekySb_CeQ
-    title: Huge Wave Slams Unlucky Surfers - Uluwatu
-    description: This was one of the biggest days of the year a few months ago. Uluwatu's
-      Outside Corner had some bombs, but the lineup looked really unpredictable.
-    thumbnails:
-      default:
-        url: https://i.ytimg.com/vi/wZYeAZtJUnE/default.jpg
-        width: 120
-        height: 90
-      medium:
-        url: https://i.ytimg.com/vi/wZYeAZtJUnE/mqdefault.jpg
-        width: 320
-        height: 180
-      high:
-        url: https://i.ytimg.com/vi/wZYeAZtJUnE/hqdefault.jpg
-        width: 480
-        height: 360
-    channelTitle: Surfers of Bali
-    liveBroadcastContent: none
-    publishTime: '2021-11-11T11:00:24Z'
-- kind: youtube#searchResult
-  etag: 6wX8jpMzMoIO_CMhTDc-JJJnfq0
-  id:
-    kind: youtube#video
-    videoId: A5oD6RhwCi0
-  snippet:
-    publishedAt: '2021-11-11T20:23:49Z'
-    channelId: UCNyxrRXa3PvyJYuFP51iU-Q
-    title: Points South with Dave Rastovich
-    description: First release of a surf trip I took with Dave Rastovich and Dan Ross
-      to explore some waves in the Southern Ocean. Featuring @John John Florence Dave
-      ...
-    thumbnails:
-      default:
-        url: https://i.ytimg.com/vi/A5oD6RhwCi0/default.jpg
-        width: 120
-        height: 90
-      medium:
-        url: https://i.ytimg.com/vi/A5oD6RhwCi0/mqdefault.jpg
-        width: 320
-        height: 180
-      high:
-        url: https://i.ytimg.com/vi/A5oD6RhwCi0/hqdefault.jpg
-        width: 480
-        height: 360
-    channelTitle: John John Florence
-    liveBroadcastContent: none
-    publishTime: '2021-11-11T20:23:49Z'
-- kind: youtube#searchResult
-  etag: VIL2GZS2JNty3DevLMg2uChlkPY
-  id:
-    kind: youtube#video
-    videoId: aS-NoUsMcac
-  snippet:
-    publishedAt: '2021-11-11T20:00:11Z'
-    channelId: UC--3c8RqSfAqYBdDjIG3UNA
-    title: Ian Walsh&#39;s How-To Guide for a Big Wave Winter | FOLLOWING Ep7
-    description: From Ian Walsh preparing his equipment for XXL waves, to giving his
-      boards their first test-drives in somewhat surprising surf, this installment
-      of Following is an ...
-    thumbnails:
-      default:
-        url: https://i.ytimg.com/vi/aS-NoUsMcac/default.jpg
-        width: 120
-        height: 90
-      medium:
-        url: https://i.ytimg.com/vi/aS-NoUsMcac/mqdefault.jpg
-        width: 320
-        height: 180
-      high:
-        url: https://i.ytimg.com/vi/aS-NoUsMcac/hqdefault.jpg
-        width: 480
-        height: 360
-    channelTitle: Red Bull Surfing
-    liveBroadcastContent: none
-    publishTime: '2021-11-11T20:00:11Z'
-- kind: youtube#searchResult
-  etag: 98euKZ4f9V9uGNzyaOUa9h103EQ
-  id:
-    kind: youtube#video
-    videoId: W7h-Yho8EB0
-  snippet:
-    publishedAt: '2019-06-15T15:21:45Z'
-    channelId: UCqhnX4jA0A5paNd1v-zEysw
-    title: 'GoPro: Top 10 Surf Moments'
-    description: 'Celebrate International Surf Day with GoPro''s Top 10 Surf Moments.
-      Shot 100% on GoPro: http://bit.ly/2wUMwfI Get stoked and subscribe: http://goo.gl/HgVXpQ
-      ...'
-    thumbnails:
-      default:
-        url: https://i.ytimg.com/vi/W7h-Yho8EB0/default.jpg
-        width: 120
-        height: 90
-      medium:
-        url: https://i.ytimg.com/vi/W7h-Yho8EB0/mqdefault.jpg
-        width: 320
-        height: 180
-      high:
-        url: https://i.ytimg.com/vi/W7h-Yho8EB0/hqdefault.jpg
-        width: 480
-        height: 360
-    channelTitle: GoPro
-    liveBroadcastContent: none
-    publishTime: '2019-06-15T15:21:45Z'
